+++ Liber/Controllers/Account Handling/RegisterViewController.swift
@@ -16,6 +16,7 @@ class RegisterViewController: UIViewController {
     // Linking text fields
     @IBOutlet var emailTextField: UITextField!
     @IBOutlet var passwordTextField: UITextField!
+    @IBOutlet var confirmPasswordTextField: UITextField!
     
     // Labels
     @IBOutlet var warningText: UILabel!
@@ -30,27 +31,23 @@ class RegisterViewController: UIViewController {
         SVProgressHUD.show()
         
         // Set up a new user on the Firebase database.
-        Auth.auth().createUser(withEmail: emailTextField.text!, password: passwordTextField.text!) { (user, error) in
-            
-            if error != nil {
-                //TODO: - If there is an error, ask the user to try again.
-                print(error!)
-                
-                self.warningText.text = "Error registering account, please try again."
-                                
-                UIView.animate(withDuration: 0.5, delay: 0, options: [.curveEaseOut], animations: {
-                    self.warningText.center.x += self.view.bounds.width
-                    self.view.layoutIfNeeded()
-                }, completion: nil)
-                
-            } else {
-                // Register successful.
-                print("Registration Successful!")
-                self.warningText.text = ""
-                
-                self.performSegue(withIdentifier: "goToGuideView", sender: self)
+        let registerParameters = ["Email": emailTextField.text!, "Password": passwordTextField.text!]
+        if conductRegistrationValidation(registerParameters: registerParameters, confirmPassword: confirmPasswordTextField.text!) {
+            registerFirebaseUser(registerParameters: registerParameters) { (result) in
+                if result {
+                    // Successful registration
+                    self.warningText.text = ""
+                    self.performSegue(withIdentifier: "goToGuideView", sender: self)
+                } else {
+                    // Error registering with Firebase.
+                    self.warningText.text = "Error registering account, please try again."
+                }
             }
-            SVProgressHUD.dismiss()
+        } else {
+            // Register validation failed, so print an error message to the user.
+            self.warningText.text = getLatestErrorMessageFromFirebaseFunctions()
         }
+        
+        SVProgressHUD.dismiss()
     }
 }